+++ internal/ent/schema/questioncategory.go
@@ -60,7 +60,7 @@ func (QuestionCategory) Annotations() []schema.Annotation {
 func (QuestionCategory) Fields() []ent.Field {
 	return []ent.Field{
 		field.String("name").NotEmpty().Comment("名称"),
-		field.Int("sort").Default(0).Comment("排序"),
+		field.Uint64("sort").Default(0).Comment("排序"),
 	}
 }
